Update To_Ruby similarly how From_Ruby was updated. #160.

+++ rice/detail/from_ruby.ipp
@@ -485,7 +485,7 @@ namespace Rice::detail
     {
       if (this->isValue_)
       {
-        return value;
+        return (unsigned long)value;
       }
       else if (value == Qnil && this->defaultValue_)
       {
